--- src/functions/copy-git-repo/requirements.txt
@@ -1,1 +1,0 @@
-git-remote-codecommit
