+++ src/supabase-studio/cr-import-repo/requirements.txt
@@ -1,0 +1,1 @@
+git-remote-codecommit
